refactored the hud modules to be in their own classes, and added a config for the positions of the modules instead of them being in hardcoded places. not localized however.
