+++ src/main/java/net/johnvictorfs/simple_utilities/helpers/StringUtils.java
@@ -1,0 +1,38 @@
+package net.johnvictorfs.simple_utilities.helpers;
+
+public class StringUtils {
+    public static String parseTime(long time) {
+        long hours = (time / 1000 + 6) % 24;
+        long minutes = (time % 1000) * 60 / 1000;
+        String ampm = "AM";
+
+        if (hours >= 12) {
+            hours -= 12;
+            ampm = "PM";
+        }
+
+        if (hours == 0) hours = 12;
+
+        String mm = "0" + minutes;
+        mm = mm.substring(mm.length() - 2);
+
+        return hours + ":" + mm + " " + ampm;
+    }
+    public static String secondsToString(int pTime) {
+        final int min = pTime / 60;
+        final int sec = pTime - (min * 60);
+
+        final String strMin = zeroPadding(min);
+        final String strSec = zeroPadding(sec);
+        return String.format("%s:%s", strMin, strSec);
+    }
+    public static String zeroPadding(int number) {
+        return (number >= 10) ? Integer.toString(number) : String.format("0%s", number);
+    }
+    public static String capitalize(String str) {
+        // Capitalize first letter of a String
+        if (str == null) return null;
+        return str.substring(0, 1).toUpperCase() + str.substring(1);
+    }
+
+}
