+++ src/main/java/net/johnvictorfs/simple_utilities/hud/elements/SimpleHudElement.java
@@ -1,0 +1,7 @@
+package net.johnvictorfs.simple_utilities.hud.elements;
+
+import me.shedaniel.math.Point;
+
+public interface SimpleHudElement {
+    void draw();
+}
